+++ NEWS.md
@@ -1,7 +1,7 @@
-# crunch 1.24.2 (under development)
-* Minor internal fixes for dataset organization.
+# crunch 1.24.3 (under development)
+* Improved robustness of API usage for moving datasets in projects.
 
-# crunch 1.24.1
+# crunch 1.24.2
 * The RStudio gadgets for `listDatasets()` and `makeArrayGadget()` have been moved to the [`crunchy`](https://github.com/Crunch-io/crunchy) package. Wiring for them, including RStudio add-in configuration, remains here, but you'll have to install that package to use them.
 * Minor fixes for backwards compatibility with the old projects API
 * Remove code paths that modify the project dataset order, which was removed from the Crunch API.